refactor: move fetching konfs into its own package
